tests/requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
firebase-admin==6.2.0
PyGithub==2.1.1
aiohttp==3.9.1 